refactor(core): move things out of repo.ts
