+++ packages/repco-core/src/mod.ts
@@ -2,7 +2,7 @@ export { repco, form } from './prisma.js'
 export * from './datasource.js'
 export * from './entity.js'
 export * from './file.js'
-export { Repo, RepoError, RelationFinder } from './repo.js'
+export { Repo, RepoError } from './repo.js'
 export * from './repo/types.js'
 export * from './shared.js'
 export type { ImportProgress } from './repo/import.js'
