+++ packages/repco-core/src/repo.ts
@@ -32,7 +32,6 @@ import {
   exportRepoToCar,
   exportRepoToCarReversed,
 } from './repo/export.js'
-import { GGraph } from './repo/graph.js'
 import { importRepoFromCar, OnProgressCallback } from './repo/import.js'
 import {
   ContentLoaderStream,
@@ -41,19 +40,18 @@ import {
 } from './repo/stream.js'
 import {
   CommitBundle,
-  CommitIpld,
   entityForm,
   HeadersForm,
   headersForm,
   RevisionIpld,
   revisionIpldToDb,
-  RootIpld,
 } from './repo/types.js'
-import { MapList } from './util/collections.js'
 import { ParseError } from './util/error.js'
-import { createEntityId, createRevisionId } from './util/id.js'
+import { createEntityId } from './util/id.js'
 import { notEmpty } from './util/misc.js'
 import { Mutex } from './util/mutex.js'
+import { RelationFinder } from './repo/relation-finder'
+import { IpldRepo } from './repo/ipld-repo'
 
 export * from './repo/types.js'
 
@@ -741,99 +739,6 @@ export class Repo {
   }
 }
 
-export class IpldRepo {
-  constructor(public record: RepoRecord, public blockstore: IpldBlockStore) {}
-  get did() {
-    return this.record.did
-  }
-  async createCommit(
-    entities: EntityInputWithHeaders[],
-    agentKeypair: ucans.EdKeypair,
-    publishingCapability: string,
-    opts: SaveBatchOpts,
-    parentCommit?: CID | null,
-  ) {
-    const agentDid = agentKeypair.did()
-    const timestamp = new Date()
-    // TODO: Reactive transaction support once timing issues with Prisma are fixed.
-    // this.blockstore = this.blockstore.transaction()
-
-    // save all revisions in blockstore
-    const revisions = []
-    for (const entity of entities) {
-      const revision = await this.createRevision(agentDid, entity)
-      if (revision) {
-        revisions.push({ ...revision, parsedContent: entity })
-      }
-    }
-
-    if (!opts.commitEmpty && !revisions.length) return null
-
-    // save commit ipld
-    const commit: CommitIpld = {
-      kind: 'commit',
-      repoDid: this.did,
-      agentDid,
-      parent: parentCommit,
-      revisions: revisions.map((r) => r.revisionCid),
-      timestamp,
-    }
-    const commitCid = await this.blockstore.put(commit)
-    const root: RootIpld = {
-      kind: 'root',
-      commit: commitCid,
-      sig: await agentKeypair.sign(commitCid.bytes),
-      cap: publishingCapability,
-      agent: agentDid,
-    }
-    const rootCid = await this.blockstore.put(root)
-
-    // batch commit ipld changes
-    this.blockstore = await this.blockstore.commit()
-
-    const bundle = {
-      root: { cid: rootCid, body: root },
-      commit: { cid: commitCid, body: commit },
-      revisions,
-    }
-    return bundle
-  }
-
-  private async createRevision(
-    agentDid: string,
-    entity: EntityInputWithHeaders,
-  ): Promise<RevisionWithUnknownContent | null> {
-    const headers = entity.headers
-    const contentCid = await this.blockstore.put(entity.content)
-    if (contentCid.toString() === entity.prevContentCid) return null
-    const id = createRevisionId()
-    if (!headers.dateModified) headers.dateModified = new Date()
-    if (!headers.dateCreated) headers.dateCreated = new Date()
-    const revisionWithoutCid: RevisionIpld = {
-      kind: 'revision',
-      id,
-      prevRevisionId: headers.prevRevisionId || null,
-      contentCid,
-      entityType: entity.type,
-      uid: entity.uid,
-      repoDid: this.did,
-      agentDid,
-      revisionUris: headers.revisionUris || [],
-      entityUris: headers.entityUris || [],
-      isDeleted: false,
-      dateModified: headers.dateModified || new Date(),
-      dateCreated: headers.dateCreated || new Date(),
-      derivedFromUid: headers.derivedFromUid,
-    }
-    const revisionCid = await this.blockstore.put(revisionWithoutCid)
-    return {
-      revision: revisionWithoutCid,
-      revisionCid,
-      content: entity.content,
-    }
-  }
-}
-
 export class RelationNotFoundError extends Error {
   constructor(public value: string, public relation: common.Relation) {
     super(
@@ -848,173 +753,6 @@ function setUid(
 ): asserts input is repco.EntityInputWithUid {
   ;(input as any).uid = uid
   input.content.uid = uid
-}
-
-export class RelationFinder {
-  prisma: Prisma.TransactionClient
-  counter = 0
-  // map of uid -> entity
-  entities: Map<string, EntityInputWithHeaders> = new Map()
-  // set not yet checked entity uris
-  pendingUris: Set<string> = new Set()
-  // map of uri -> Relation[]
-  relsByUri: MapList<common.Relation> = new MapList()
-  // set of uris checked and found missing
-  missingUris: Set<string> = new Set()
-  // map of uri -> uid
-  uriMap: Map<string, string> = new Map()
-
-  static resolve(repo: Repo, entities: EntityInputWithHeaders[]) {
-    const resolver = new RelationFinder(repo)
-    resolver.pushBatch(entities)
-    return resolver.resolve()
-  }
-
-  static resolveLinks(repo: Repo, links: common.Link[]) {
-    const resolver = new RelationFinder(repo)
-    for (const link of links) {
-      resolver.pushLink(link)
-    }
-    return resolver.resolve()
-  }
-
-  constructor(public repo: Repo) {
-    this.prisma = repo.prisma
-  }
-
-  pushLink(value: common.Link, relation?: common.Relation) {
-    // TODO: Throw if uri missing?
-    if (value.uid || !value.uri || this.missingUris.has(value.uri)) return
-    // check if already in map: resolve now
-    if (this.uriMap.has(value.uri)) {
-      value.uid = this.uriMap.get(value.uri)
-    } else {
-      this.pendingUris.add(value.uri)
-      if (relation) this.relsByUri.push(value.uri, relation)
-    }
-  }
-
-  pushEntity(entity: EntityInputWithHeaders) {
-    const uid = entity.uid
-    if (this.entities.has(uid)) return
-    this.entities.set(uid, entity)
-    if (entity.headers.entityUris) {
-      for (const uri of entity.headers.entityUris) {
-        this.discoveredUid(uri, uid)
-      }
-    }
-    const relations = repco.extractRelations({ ...entity, uid })
-    for (const relation of relations) {
-      for (const value of relation.values) {
-        this.pushLink(value, relation)
-      }
-    }
-  }
-
-  pushBatch(entities: EntityInputWithHeaders[]) {
-    entities.forEach((entity) => this.pushEntity(entity))
-  }
-
-  setLinkUid(
-    entity: EntityInputWithHeaders,
-    relation: common.Relation,
-    uri: string,
-    uid: string,
-  ) {
-    const { field, multiple } = relation
-    const entityAny = entity as any
-    if (multiple) {
-      for (const link of entityAny.content[field] as common.Link[]) {
-        if (link.uri === uri) link.uid = uid
-      }
-    } else {
-      entityAny.content[field] = { uid, uri }
-    }
-  }
-
-  discoveredUid(uri: string, uid: string) {
-    this.uriMap.set(uri, uid)
-    const relations = this.relsByUri.get(uri)
-    if (!relations || !relations.length) return
-    for (const relation of relations) {
-      const entity = this.entities.get(relation.uid)
-      if (entity) this.setLinkUid(entity, relation, uri, uid)
-    }
-    this.pendingUris.delete(uri)
-  }
-
-  getByUri(uri: string) {
-    const uid = this.uriMap.get(uri)
-    if (!uid) return
-    return this.entities.get(uid)
-  }
-
-  async resolve() {
-    // TODO: Max iterations
-    while (this.pendingUris.size) {
-      // create set of all uri relations
-      // try to fetch them locally from the revision table
-      const res = await this.prisma.revision.findMany({
-        where: {
-          entityUris: { hasSome: [...this.pendingUris] },
-        },
-        select: { uid: true, entityUris: true },
-      })
-      for (const row of res) {
-        for (const uri of row.entityUris) {
-          this.discoveredUid(uri, row.uid)
-        }
-      }
-
-      // try to fetch them from the datasources
-      const { fetched, notFound } = await this.repo.dsr.fetchEntities(
-        this.repo,
-        [...this.pendingUris],
-      )
-
-      notFound.forEach((uri) => {
-        this.missingUris.add(uri)
-        this.pendingUris.delete(uri)
-      })
-      const existingEntities = []
-      const newEntityInputs = []
-      for (const input of fetched) {
-        let existing = false
-        if (input.entityUris) {
-          // check if the entity was already fetched in this resolver session
-          for (const uri of input.entityUris) {
-            const entity = this.getByUri(uri)
-            if (entity) {
-              existingEntities.push(entity)
-              existing = true
-              break
-            }
-          }
-        }
-        if (!existing) newEntityInputs.push(input)
-      }
-      const newEntities = await this.repo.assignUids(
-        parseEntities(newEntityInputs),
-      )
-      this.pushBatch([...existingEntities, ...newEntities])
-    }
-
-    // Sort the results. This also throws for ciruclar references.
-    const graph = new GGraph()
-    for (const entity of this.entities.values()) {
-      const edges = repco
-        .extractRelations(entity)
-        .map((r) => r.values)
-        .flat()
-        .filter((x) => x.uid && this.entities.has(x.uid))
-        .map((x) => x.uid) as string[]
-      graph.push(entity.uid, edges)
-    }
-    const stack = graph.resolve()
-
-    // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-    return stack.map((uid) => this.entities.get(uid)!)
-  }
 }
 
 function assertFullClient(
@@ -1043,11 +781,6 @@ function parseEntity(input: unknown): EntityFormWithHeaders {
   }
 }
 
-function parseEntities(inputs: unknown[]): EntityFormWithHeaders[] {
-  // eslint-disable-next-line no-useless-catch
-  try {
-    return inputs.map(parseEntity)
-  } catch (err) {
-    throw err
-  }
+export function parseEntities(inputs: unknown[]): EntityFormWithHeaders[] {
+  return inputs.map(parseEntity)
 }
